current work with bug on real time mode mqtt

+++ Mikrocontroller/BO_Tracker/lib/Struct/TrackerMode.h
@@ -12,7 +12,7 @@ struct TrackerModes {
     bool GeoFenMode = false;
     bool RequestMode = true;
     bool Modem_Off = false;
-    unsigned int period = 5000UL;
+    unsigned long period = 5000UL;
     // GeoFencing
     float geoLatitude = 0;
     float geoLongitude = 0;
